clippy

+++ src/game/gcl.rs
@@ -46,5 +46,5 @@ pub fn progress_total() -> f64 {
 pub fn total_for_level(level: u32) -> f64 {
     // formula from
     // https://github.com/screeps/engine/blob/6d498f2f0db4e0744fa6bf8563836d36b49b6a29/src/game/game.js#L117
-    ((level - 1) as f64).powf(GCL_POW as f64) * GCL_MULTIPLY as f64
+    ((level - 1) as f64).powf(GCL_POW) * GCL_MULTIPLY as f64
 }
